Move CaptionAttention to attention.py
